+++ CHANGELOG.md
@@ -124,6 +124,8 @@
 102. HarmonyShield now protects Methods and Assembly from being patched.
 103. Added HarmonyShield Protection to MelonLoader, SM_Il2Cpp, SM_Mono, SM_Mono.Pre-2017, and SM_Mono.Pre-5.
 104. Fixed Issue with Unneeded Return in Logger breaking Logs after Plugin Loading.
+105. Fixed Issue with Console Buffer.
+106. Fixed Issue with Console Coloring.
 
 ---